Refactor AcqTable base and get_stars to core module
